Fix/image issue (#20)
* fix: Update image source URLs for product and offer pages

* feat: Add profile image upload functionality

The code changes in the "Settings" page component add the functionality to upload a profile image. Users can now select an image file and upload it to their profile. The profile image is displayed in a circular container, and users can click on it to open the file selection dialog. This enhancement improves the user experience and allows users to personalize their profiles.

Co-authored-by: Adilsagheer <adilsagheer051@gmail.com>

* refactor: Remove commented code for gallery image upload

The code changes in the "EditProduct" component remove the commented code that was previously used for uploading gallery images. This code is no longer needed and can be safely removed. This refactor improves code cleanliness and reduces confusion for future developers.

Co-authored-by: Adilsagheer <adilsagheer051@gmail.com>

* refactor: Remove unnecessary commented code for image upload

* fix: Update email address in ContactSupport page

The code changes in the ContactSupport page update the email address displayed to users. The previous email address was incorrect, and it has been replaced with the correct email address "www.grazle.co.in". This fix ensures that users have the correct contact information when using the ContactSupport feature.

Co-authored-by: Adilsagheer <adilsagheer051@gmail.com>

---------

Co-authored-by: Adilsagheer <adilsagheer051@gmail.com>